+++ synapse_core/__init__.py
@@ -1,19 +1,21 @@
-__version__ = "0.7.0"
+__version__ = "1.0.0"
 
 from .exceptions import CollectionNotFoundError, SourceNotFoundError, SynapseError, TableNotFoundError
 from .logger import setup_logging
 from .models import IngestProgress, IngestResult, PurgeResult, QueryResult
-from .pipeline import ingest, ingest_many, purge, query, reset, sources
+from .pipeline import ingest, ingest_async, ingest_many, purge, query, query_async, reset, sources
 from .sqlite_ingester import ingest_sqlite
 
 __all__ = [
     "__version__",
     # core API
     "ingest",
+    "ingest_async",
     "ingest_many",
     "ingest_sqlite",
     "purge",
     "query",
+    "query_async",
     "reset",
     "setup_logging",
     "sources",